backend/requirements.txt
hypercorn
fastapi
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
bcrypt
python-jose[cryptography]
pydantic[email]
python-dotenv
python-multipart
schwifty
greenlet
apscheduler
fastapi-mail
contextvars
python-dateutil
pytest
pytest-asyncio
pytest-cov
httpx
aiosqlite
fints
scikit-learn
fpdf2